% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband_map}
\alias{cytoband_map}
\title{Information about cytobands in 5 most recent assemblies}
\format{List of 5, each list item is a data.table with 5 columns: chromosome,
  cytoband start, cytoband end, cytoband name, and stain value
\describe{
  \item{hg16}{data.table with 5 columns}
  \item{...}{data.table with 5 columns}
}}
\source{
\url{https://hgdownload.cse.ucsc.edu/goldenPath/hg19/database/cytoBand.txt.gz}
}
\usage{
cytoband_map
}
\description{
List of 5, each list item is a data.table with 5 columns: chromosome,
  cytoband start, cytoband end, cytoband name, and stain value.
  Used by the coord_to_arm function.
}
\keyword{datasets}

