% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_map.R
\name{dt_to_385k}
\alias{dt_to_385k}
\title{Converts aCGH data to hg18 385k design}
\usage{
dt_to_385k(dt, assembly = "hg18")
}
\arguments{
\item{dt}{data.table with aCGH data. First two colums should be names chrom and pos
and contain information about the location of input probes. Chromosomes should be
names "chr1", "chrY", etc.
The remaining columns are interpreted as samples with the column name denoting
the sample name.}

\item{assembly}{String specifying assembly of input data. Can take values "hg18"
(default) or "hg17" (in which case, design must match hg17 385k design).}
}
\value{
data.table with mapped data. First two colums are chrom and pos
  and contain information about the location of input probes. The remaining columns
  are mapped data for each sample.
}
\description{
Converts aCGH data to a hg18 385k design.
  Each input probe is mapped to its nearest output probe in the 385k design.
  If multiple inputs match to a single output, their mean is computed. If no input
  matches to a 385k probe, an NA value is returned for that output.
}

