% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg17_hg18_map}
\alias{hg17_hg18_map}
\title{Map from probes in hg17 385k array design to locations in hg18 assembly}
\format{A data table with 388560 obs. of 4 variables:
\describe{
  \item{hg17chrom}{chromosome of probe in hg17, in the format "chr1", "chrY", etc.}
  \item{hg17pos}{position of probe in hg17, position of probe on chromosome}
  \item{hg18chrom}{chromosome of corresponding location in hg17, in the format "chr1", "chrY", etc.}
  \item{hg18pos}{position of corresponding location in hg17, on chromosome}
}}
\source{
\url{https://genome.ucsc.edu/cgi-bin/hgLiftOver}
}
\usage{
hg17_hg18_map
}
\description{
A data.table with genomic coordinates of 388560 probes in an array,
  using hg17 assembly. Each probe has listed the corresponding genomic
  coordinates in hg18 assembly.
}
\keyword{datasets}

