% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hg18_385k_pos}
\alias{hg18_385k_pos}
\title{Location of probes in hg18 385k design array}
\format{A data table with 385815 obs. of 2 variables:
\describe{
  \item{chrom}{chromosome, in the format "chr1", "chrY", etc.}
  \item{pos}{position, position of probe on chromosome}
}}
\source{
Desgin files
}
\usage{
hg18_385k_pos
}
\description{
A data.table with genomic coordinates of 385815 probes in an array,
  using hg18 assembly. Extracted from design files.
}
\keyword{datasets}

