% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcsmooth.R
\name{kcmakefilter}
\alias{kcmakefilter}
\title{Create index from sig_region object}
\usage{
kcmakefilter(sig_regions)
}
\arguments{
\item{sig_regions}{A sig_regions object from kcsmart}
}
\value{
data.table with 4 columns: number of gain/loss, chrom(osome) of region, whether its a gain or loss (pos/neg)
  and the num(ber) of kcprobe (genomic location of kcprobes is just sampleDensity*(number-1))
}
\description{
Function to extract information from a sig_region object, to a list of all kcprobes within
those regions
}

