% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kcsmooth.R
\name{kcsmooth}
\alias{kcsmooth}
\title{Smooth samples with KCsmart algorithm}
\usage{
kcsmooth(dt, mirrorLocs, sigma = 1e+06, sampleDensity = 50000,
  maxmem = 1000, verbose = T, what = "both")
}
\arguments{
\item{dt}{data.table with aCGH data. First two colums should be names chrom and pos
and contain information about the location of input probes. Chromosomes should be
names "1", "Y", etc.}

\item{mirrorLocs}{List containing the chromosome start, centromere and end positions}

\item{sigma}{The kernel width}

\item{sampleDensity}{The sample point matrix resolution}

\item{maxmem}{This parameter controls memory usage, set to lower value to lower memory consumption}

\item{verbose}{If set to false, no progress information is displayed}

\item{what}{character, determines what data should be returned. "pos" for only positive KCscore,
"neg" for only negative KCscore, "both" for both positive and negative KCscore.}
}
\value{
data.table with smoothed values. First column is chromosome, second column indicates if value is from positive
  or negative smoothed values. All other columns are smoothed values for each sample.
}
\description{
Wrapper to smooth aCGH data with algorithm from KCsmart.
}

