% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assembly_map.R
\name{probe_to_385k}
\alias{probe_to_385k}
\title{Finds nearest probe on hg18_385k}
\usage{
probe_to_385k(x, chrom)
}
\arguments{
\item{x}{Numeric vector, with genomic coordinate of probe}

\item{chrom}{Character or character vector (in the form "chr11", "chrX", etc.),
with chromosome of probe. All elements must be identical.}
}
\value{
Numeric vector, with genomic coordinate of closest probe in the hg18_385k design
}
\description{
Returns nearest probe on a 385k array design in hg18 assembly
}
\examples{
  # Create sample data
  library(data.table)
  dat <- data.table(chrom = c("chr1", "chr1"), pos = c(46348, 46349))
  dat[, pos_385k := map_to_385k(pos, chrom), by = chrom]
  dat
}

