% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waai_caai.R
\name{waai}
\alias{waai}
\title{Whole-arm aberration index (WAAI)}
\usage{
waai(seg_values, seg_nprobes)
}
\arguments{
\item{seg_values}{Numeric vector, with mean (or other measure) of log ratio for each segment on chromosome arm.}

\item{seg_nprobes}{Numeric vector, with number of probes in each segment.}
}
\value{
Whole-arm abberation index for the given chromosome arm.
}
\description{
Computes whole-arm aberration index (WAAI) for a chromosome arm. WAAI is an indicator of whole-arm gains and losses.
  See Russnes et al (DOI: 10.1126/scitranslmed.3000611) for a detailed definition.
}
\examples{
# Read aCGH data to data.frame dat
seg <- copynumber::pcf(dat)
seg.dt <- data.table::data.table(seg)
seg.dt[,.(waai = waai(mean, n.probes)), by=.(chrom, arm)]
}

